implementation/requirements.txt
astroid==2.5.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
psycopg2-binary==2.8.6
pylint==2.7.2
scapy==2.4.4
toml==0.10.2
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.6.3
click>=7.1.2